context less version of Trans component
