+++ src/context.js
@@ -1,30 +1,11 @@
 import { createContext } from 'react';
-import { unescape } from './unescape';
+import { getDefaults, setDefaults } from './defaults';
 
-let defaultOptions = {
-  bindI18n: 'languageChanged',
-  bindI18nStore: '',
-  // nsMode: 'fallback' // loop through all namespaces given to hook, HOC, render prop for key lookup
-  transEmptyNodeValue: '',
-  transSupportBasicHtmlNodes: true,
-  transWrapTextNodes: '',
-  transKeepBasicHtmlNodesFor: ['br', 'strong', 'i', 'p'],
-  // hashTransKey: key => key // calculate a key for Trans component based on defaultValue
-  useSuspense: true,
-  unescape,
-};
+export { getDefaults, setDefaults };
 
 let i18nInstance;
 
 export const I18nContext = createContext();
-
-export function setDefaults(options = {}) {
-  defaultOptions = { ...defaultOptions, ...options };
-}
-
-export function getDefaults() {
-  return defaultOptions;
-}
 
 export class ReportNamespaces {
   constructor() {
